requirements.txt
Pillow==10.0.1
Pillow==11.1.0
